N.O.T.A/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml_html_clean==0.4.3
trafilatura==1.12.1
aiosqlite==0.21.0
httpx==0.27.0